apply clipping removed from clip plane dialog. captions for surface meshes can be fixed to screen position or moved with the attached mesh.

+++ drishti/trisetinformation.h
@@ -49,7 +49,7 @@ class TrisetInformation
   QColor captionColor;
   QFont captionFont;
   Vec captionPosition;
-  int cpDx, cpDy;
+  float cpDx, cpDy;
   float opacity;
 };
